+++ lib/main.dart
@@ -1,0 +1,2 @@
+export "./shammo.dart";
+export "./utils.dart";