+++ lib/shammo.dart
@@ -1,21 +1,22 @@
 library shammo;
+
 import 'dart:async';
 
 import 'package:flutter/material.dart';
 import 'package:shammo/utils.dart';
 
 class Badge extends StatelessWidget {
-  Badge({Key key, this.child, this.badge, this.badgeColor}) : super(key: key);
-  final Widget child;
-  final Widget badge;
-  final Color badgeColor;
+  Badge({Key? key, this.child, this.badge, this.badgeColor}) : super(key: key);
+  final Widget? child;
+  final Widget? badge;
+  final Color? badgeColor;
 
   Widget build(BuildContext context) {
     return Container(
       margin: EdgeInsets.only(right: 10),
       child: Stack(
         children: [
-          child,
+          child!,
           Positioned(
             child: Container(
               child: Center(child: badge),
@@ -35,12 +36,18 @@ class Badge extends StatelessWidget {
 }
 
 class ContainerButton extends StatelessWidget {
-  ContainerButton({Key key, this.body, this.action, this.bodyColor, this.margin, this.padding})
+  ContainerButton(
+      {Key? key,
+      this.body,
+      this.action,
+      this.bodyColor,
+      this.margin,
+      this.padding})
       : super(key: key);
-  final Widget body;
-  final Widget action;
-  final Color bodyColor;
-  final EdgeInsets margin, padding;
+  final Widget? body;
+  final Widget? action;
+  final Color? bodyColor;
+  final EdgeInsets? margin, padding;
 
   @override
   Widget build(BuildContext context) {
@@ -68,14 +75,15 @@ Widget dropDropButtonUnderline(BuildContext context) {
     height: 40,
     padding: EdgeInsets.all(10),
     width: width(context),
-    decoration:
-        BoxDecoration(borderRadius: BorderRadius.circular(10), border: Border.all(color: Colors.grey)),
+    decoration: BoxDecoration(
+        borderRadius: BorderRadius.circular(10),
+        border: Border.all(color: Colors.grey)),
   );
 }
 
 class PostCard extends StatelessWidget {
   PostCard(
-      {Key key,
+      {Key? key,
       this.title,
       this.actions,
       this.action,
@@ -88,21 +96,20 @@ class PostCard extends StatelessWidget {
       this.avatarDecoration,
       this.padding})
       : super(key: key);
-  final Widget title;
-  final List<Widget> actions;
-  final Widget action;
-  final Widget subheader;
-  final Widget avatar;
-  final Widget content;
-  final BoxDecoration decoration;
-  final BoxConstraints constraints;
-  final EdgeInsets margin;
-  final EdgeInsets padding;
+  final Widget? title;
+  final List<Widget>? actions;
+  final Widget? action;
+  final Widget? subheader;
+  final Widget? avatar;
+  final Widget? content;
+  final BoxDecoration? decoration;
+  final BoxConstraints? constraints;
+  final EdgeInsets? margin;
+  final EdgeInsets? padding;
   final avatarDecoration;
 
   @override
   Widget build(BuildContext context) {
-    
     return Container(
       decoration: decoration,
       constraints: constraints,
@@ -116,23 +123,26 @@ class PostCard extends StatelessWidget {
             child: Row(
               children: [
                 if (this.avatar != null)
-                      Container(
-                          margin: EdgeInsets.only(right: 10),
-                          decoration: avatarDecoration,
-                          child: this.avatar),
-                if (this.title != null) Expanded(flex: 9, child: this.title),
+                  Container(
+                      margin: EdgeInsets.only(right: 10),
+                      decoration: avatarDecoration,
+                      child: this.avatar),
+                if (this.title != null) Expanded(flex: 9, child: this.title!),
                 if (this.action != null) Container(child: this.action),
               ],
             ),
           ),
           if (this.subheader != null)
-            Container(margin: EdgeInsets.all(10), width: width(context), child: this.subheader),
+            Container(
+                margin: EdgeInsets.all(10),
+                width: width(context),
+                child: this.subheader),
           if (this.content != null)
             Container(
               margin: EdgeInsets.all(5),
               width: width(context),
               child: Column(
-                children: [this.content],
+                children: [this.content!],
               ),
             ),
           if (this.actions != null)
@@ -140,7 +150,7 @@ class PostCard extends StatelessWidget {
               margin: EdgeInsets.all(5),
               width: width(context),
               child: Row(
-                children: this.actions,
+                children: this.actions!,
               ),
             ),
         ],
@@ -151,9 +161,9 @@ class PostCard extends StatelessWidget {
 
 class Carousel extends StatefulWidget {
   Carousel(
-      {Key key,
-      @required this.items,
-      @required this.aspectRatio,
+      {Key? key,
+      required this.items,
+      required this.aspectRatio,
       this.indicator,
       this.indicatorColor,
       this.activeIndicatorWidth,
@@ -166,11 +176,15 @@ class Carousel extends StatefulWidget {
       : super(key: key);
   final List<Widget> items;
   final double aspectRatio;
-  final Widget indicator;
-  final Color indicatorColor;
-  final double activeIndicatorWidth, indicatorWidth, activeIndicatorHeight, indicatorHeight, indicatorRadius;
-  final GestureTapCallback onTap;
-  final ValueChanged<int> onPageChange;
+  final Widget? indicator;
+  final Color? indicatorColor;
+  final double? activeIndicatorWidth,
+      indicatorWidth,
+      activeIndicatorHeight,
+      indicatorHeight,
+      indicatorRadius;
+  final GestureTapCallback? onTap;
+  final ValueChanged<int>? onPageChange;
   _CarouselState createState() => _CarouselState();
 }
 
@@ -178,7 +192,7 @@ class _CarouselState extends State<Carousel> {
   final PageController controller = PageController();
   int page = 0;
 
-  double setIndicatorWidth(int index) {
+  double? setIndicatorWidth(int index) {
     if (index == page) {
       if (widget.activeIndicatorWidth != null) {
         return widget.activeIndicatorWidth;
@@ -194,7 +208,7 @@ class _CarouselState extends State<Carousel> {
     }
   }
 
-  double setIndicatorHeight() {
+  double? setIndicatorHeight() {
     if (widget.indicatorHeight != null) {
       return widget.indicatorHeight;
     } else {
@@ -214,7 +228,8 @@ class _CarouselState extends State<Carousel> {
               controller: controller,
               children: widget.items,
               onPageChanged: (pageIndex) {
-                if (widget.onPageChange != null) widget.onPageChange(pageIndex);
+                if (widget.onPageChange != null)
+                  widget.onPageChange!(pageIndex);
                 setState(() {
                   page = pageIndex;
                 });
@@ -226,7 +241,8 @@ class _CarouselState extends State<Carousel> {
               width: width(context),
               alignment: Alignment.center,
               height: setIndicatorHeight(),
-              child: Row(mainAxisAlignment: MainAxisAlignment.center, children: [
+              child:
+                  Row(mainAxisAlignment: MainAxisAlignment.center, children: [
                 for (int index = 0; index < widget.items.length; index++)
                   Container(
                     width: setIndicatorWidth(index),
@@ -234,8 +250,12 @@ class _CarouselState extends State<Carousel> {
                     margin: EdgeInsets.only(right: 5),
                     decoration: BoxDecoration(
                       borderRadius: BorderRadius.circular(
-                          widget.indicatorRadius != null ? widget.indicatorRadius : 10),
-                      color: widget.indicatorColor == null ? Colors.blueAccent : widget.indicatorColor,
+                          widget.indicatorRadius != null
+                              ? widget.indicatorRadius!
+                              : 10),
+                      color: widget.indicatorColor == null
+                          ? Colors.blueAccent
+                          : widget.indicatorColor,
                     ),
                   )
               ]))
@@ -245,21 +265,20 @@ class _CarouselState extends State<Carousel> {
   }
 }
 
-
 class MaterialRadio extends StatefulWidget {
   const MaterialRadio(
-      {Key key,
+      {Key? key,
       this.color,
       this.selectColor,
       this.radius,
-      @required this.onChange,
+      required this.onChange,
       this.size,
       this.disenabled})
       : super(key: key);
-  final Color color;
-  final Color selectColor;
-  final double size, radius;
-  final bool disenabled;
+  final Color? color;
+  final Color? selectColor;
+  final double? size, radius;
+  final bool? disenabled;
   final void Function(bool) onChange;
   @override
   _MaterialRadioState createState() => _MaterialRadioState();
@@ -278,7 +297,7 @@ class _MaterialRadioState extends State<MaterialRadio> {
           if (widget.disenabled == null || (widget.disenabled == false)) {
             selected ? selected = false : selected = true;
           }
-          if (widget.onChange != null) widget.onChange(selected);
+        widget.onChange(selected);
         });
       },
       child: Stack(
@@ -295,8 +314,8 @@ class _MaterialRadioState extends State<MaterialRadio> {
           ),
           if (selected)
             Container(
-              width: widget.size != null ? widget.size / 2 : 12.5,
-              height: widget.size != null ? widget.size / 2 : 12.5,
+              width: widget.size != null ? widget.size! / 2 : 12.5,
+              height: widget.size != null ? widget.size! / 2 : 12.5,
               decoration: BoxDecoration(
                   color: widget.selectColor ?? Colors.white,
                   borderRadius: BorderRadius.circular(
@@ -310,18 +329,18 @@ class _MaterialRadioState extends State<MaterialRadio> {
 }
 
 class CountDown extends StatefulWidget {
-  const CountDown({Key key, this.style, @required this.duration})
+  const CountDown({Key? key, this.style, required this.duration})
       : super(key: key);
-  final TextStyle style;
+  final TextStyle? style;
   final int duration;
   @override
   _CountDownState createState() => _CountDownState();
 }
 
 class _CountDownState extends State<CountDown> {
-  int interval;
+  late int interval;
   String time = "";
-  Timer timer;
+  late Timer timer;
 
   @override
   void initState() {
@@ -357,23 +376,22 @@ class _CountDownState extends State<CountDown> {
   }
 }
 
-
 class MenuButton extends StatelessWidget {
   const MenuButton(
-      {@required this.header,
-      @required this.icon,
+      {required this.header,
+      required this.icon,
       this.description,
       this.color,
-      @required this.route});
+      required this.route});
   final String header;
-  final String description;
+  final String? description;
   final IconData icon;
   final Widget route;
-  final Color color;
+  final Color? color;
   @override
   Widget build(BuildContext context) {
     return Container(
-        color: color??Colors.blueAccent,
+        color: color ?? Colors.blueAccent,
         padding: const EdgeInsets.symmetric(vertical: 20, horizontal: 10),
         margin: const EdgeInsets.symmetric(horizontal: 30)
             .add(const EdgeInsets.only(top: 10)),
@@ -382,7 +400,10 @@ class MenuButton extends StatelessWidget {
           child: Row(
             mainAxisAlignment: MainAxisAlignment.spaceBetween,
             children: [
-              Icon(icon, color: color??Colors.blueAccent,),
+              Icon(
+                icon,
+                color: color ?? Colors.blueAccent,
+              ),
               Container(
                 margin: const EdgeInsets.symmetric(horizontal: 5),
                 width: width(context) / 2,
@@ -398,35 +419,38 @@ class MenuButton extends StatelessWidget {
                       Container(
                           width: (width(context) / 2) - 5,
                           child: Text(
-                            description,
+                            description!,
                             style: const TextStyle(fontSize: 11),
                           ))
                   ],
                 ),
               ),
               RotatedBox(
-                  quarterTurns: 2, child: Icon(Icons.arrow_back_ios,color: color??Colors.blueAccent))
+                  quarterTurns: 2,
+                  child: Icon(Icons.arrow_back_ios,
+                      color: color ?? Colors.blueAccent))
             ],
           ),
         ));
   }
 }
 
-
 class CustomTextField extends StatelessWidget {
   const CustomTextField(
-      {@required this.label,
-      @required this.controller,
+      {required this.label,
+      required this.controller,
       this.prefix,
       this.suffix,
-      this.hint,this.color, this.border});
+      this.hint,
+      this.color,
+      this.border});
   final String label;
   final TextEditingController controller;
-  final Widget suffix;
-  final Widget prefix;
-  final String hint;
-  final Color color;
-  final InputBorder border;
+  final Widget? suffix;
+  final Widget? prefix;
+  final String? hint;
+  final Color? color;
+  final InputBorder? border;
   @override
   Widget build(BuildContext context) {
     return Column(
@@ -441,10 +465,11 @@ class CustomTextField extends StatelessWidget {
             controller: controller,
             keyboardType: TextInputType.number,
             decoration: InputDecoration(
-              border: border,
+                border: border,
                 hintText: hint,
-                hintStyle: const TextStyle(fontSize: 16,fontWeight: FontWeight.w600),
-                suffixStyle: TextStyle(color:color),
+                hintStyle:
+                    const TextStyle(fontSize: 16, fontWeight: FontWeight.w600),
+                suffixStyle: TextStyle(color: color),
                 prefixStyle: TextStyle(color: color),
                 prefixIcon: prefix,
                 suffixIcon: suffix),
@@ -453,4 +478,120 @@ class CustomTextField extends StatelessWidget {
       ],
     );
   }
+}
+
+class MaterialCard extends StatelessWidget {
+  const MaterialCard(
+      {Key? key,
+      this.title,
+      this.actions,
+      this.action,
+      this.subheader,
+      this.avatar,
+      this.content,
+      this.decoration,
+      this.constraints,
+      this.margin,
+      this.avatarDecoration,
+      this.padding})
+      : super(key: key);
+  final Widget? title;
+  final List<Widget>? actions;
+  final Widget? action;
+  final Widget? subheader;
+  final Widget? avatar;
+  final Widget? content;
+  final BoxDecoration? decoration;
+  final BoxConstraints? constraints;
+  final EdgeInsets? margin;
+  final EdgeInsets? padding;
+  final BoxDecoration? avatarDecoration;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      decoration: decoration,
+      constraints: constraints,
+      padding: padding,
+      margin: margin,
+      child: Column(
+        children: [
+          Container(
+            margin: const EdgeInsets.all(5),
+            width: width(context),
+            child: Row(
+              children: [
+                if (avatar != null)
+                  Container(
+                      child: Container(
+                          margin: const EdgeInsets.only(right: 10),
+                          decoration: avatarDecoration,
+                          child: avatar)),
+                if (title != null) Expanded(flex: 9, child: title!),
+                if (action != null) Container(child: action),
+              ],
+            ),
+          ),
+          if (subheader != null)
+            Container(
+                margin: const EdgeInsets.all(10),
+                width: width(context),
+                child: subheader),
+          if (content != null)
+            Container(
+              margin: const EdgeInsets.all(5),
+              width: width(context),
+              child: Column(
+                children: [content!],
+              ),
+            ),
+          if (actions != null)
+            Container(
+              margin: const EdgeInsets.all(5),
+              width: width(context),
+              child: Row(
+                children: actions!,
+              ),
+            ),
+        ],
+      ),
+    );
+  }
+}
+
+class SnackbarButton extends StatelessWidget {
+  const SnackbarButton(
+      {required this.onPressed,
+      this.child,
+      this.color,
+      this.textColor,
+      this.text});
+  final Widget? child;
+  final Color? color, textColor;
+  final String? text;
+  final Function() onPressed;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      width: width(context),
+      child: ElevatedButton(
+        style: ElevatedButton.styleFrom(primary: color ?? Colors.blueAccent),
+        onPressed: onPressed,
+        child: child ??
+            Text(
+              text!,
+              style: TextStyle(color: textColor ?? Colors.white),
+            ),
+      ),
+    );
+  }
+}
+
+ThemeData getTheme(BuildContext context) {
+  return Theme.of(context);
+}
+
+MediaQueryData getMedia(BuildContext context) {
+  return MediaQuery.of(context);
 }