+++ CHANGELOG.md
@@ -1,3 +1,15 @@
-## 0.0.1
+## 1.0.0
 
-* TODO: Describe initial release.
+* This is the initial release
+
+## 2.0.0
+
+* Support for Null Safety added
+
+## 2.0.1
+
+* The MaterialRadio, MaterialCard, SnackbarButton were added and getTheme and getMedia were added
+
+## 3.0.0
+ * commaSeperatedNumber, toDouble, toInt, copy, convertToColor, getColorCode, confirmAction, showSnackBar were added
+ * showsnackbar no longer inherits the canvas color